backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.5.2
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
google-genai==1.0.0
apscheduler==3.10.4
python-dotenv==1.0.0
slowapi==0.1.9
RestrictedPython==7.0
starlette==0.38.2
python-multipart==0.0.9
